turn cli into lib
